+++ src/main/java/com/management/documenttype/repository/DocumentTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentTypeRepository extends JpaRepository<DocumentType, Integer> {
+public interface DocumentTypeRepository extends JpaRepository<DocumentType, Long> {
 }
