+++ src/main/java/com/management/projectoccurrence/controller/ProjectOccurrenceController.java
@@ -32,25 +32,25 @@ public class ProjectOccurrenceController {
 
     @GetMapping("/occurrences/{id}")
     @Operation(summary = "Finds an occurrence by ID")
-    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Integer id) {
+    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Long id) {
         return ResponseEntity.ok(projectOccurrenceService.findById(id));
     }
 
     @GetMapping("/projects/{projectId}/occurrences")
     @Operation(summary = "Finds all occurrences for a project")
-    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Integer projectId) {
+    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Long projectId) {
         return ResponseEntity.ok(projectOccurrenceService.listByProject(projectId));
     }
 
     @PutMapping("/occurrences/{id}")
     @Operation(summary = "Updates an occurrence")
-    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Integer id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
+    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Long id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
         return ResponseEntity.ok(projectOccurrenceService.update(id, request));
     }
 
     @DeleteMapping("/occurrences/{id}")
     @Operation(summary = "Deletes an occurrence")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         projectOccurrenceService.delete(id);
         return ResponseEntity.noContent().build();
     }
